Idan CR: bugfixes. tests are passing
