+++ tests/unit/monitor/alerts/group_alerts_by_table/mock_data.py
@@ -11,9 +11,9 @@ MODEL_3 = "models.bla.model3"
 OWNER_1 = "owner1@elementary-data.com"
 OWNER_2 = "owner2@elementary-data.com"
 OWNER_3 = "owner3@elementary-data.com"
-TAGS_1 = "marketing, finance"
-TAGS_2 = "finance, data-ops"
-TAGS_3 = "marketing, finance, data-ops"
+TAGS_1 = ["marketing", "finance"]
+TAGS_2 = ["finance", "data-ops"]
+TAGS_3 = ["#marketing", "#finance", "#data-ops"]
 DETECTED_AT = "1992-11-11 20:00:03+0200"
 EMPTY_DICT = json.dumps(dict())
 OTHER_CHANNEL_IN_DICT = json.dumps({CHANNEL_KEY: OTHER_CHANNEL})
